Removed problem3 Student class and completed problem5 Student class.
