+++ .idea/workspace.xml
@@ -197,10 +197,10 @@
       <screen x="0" y="0" width="1600" height="900" />
     </state>
     <state x="608" y="137" width="374" height="633" key="#com.intellij.ide.util.MemberChooser/0.0.1600.900@0.0.1600.900" timestamp="1585305741085" />
-    <state x="346" y="86" width="1022" height="814" key="CommitChangelistDialog2" timestamp="1585305439488">
+    <state x="346" y="86" width="1022" height="814" key="CommitChangelistDialog2" timestamp="1585305914300">
       <screen x="0" y="0" width="1600" height="900" />
     </state>
-    <state x="346" y="86" width="1022" height="814" key="CommitChangelistDialog2/0.0.1600.900@0.0.1600.900" timestamp="1585305439488" />
+    <state x="346" y="86" width="1022" height="814" key="CommitChangelistDialog2/0.0.1600.900@0.0.1600.900" timestamp="1585305914300" />
     <state width="1554" height="133" key="GridCell.Tab.0.bottom" timestamp="1585240160476">
       <screen x="0" y="0" width="1600" height="900" />
     </state>
